rewrote VirtualDate and adds types for `date.ts`
There is an issue matching the type signature of VirtualDate to the native Date, due to the unconventional aspect of `Date()` returning a string while `new Date(...)` returns a Date. TypeScript does not seem to be able to distinguish each particular case (e.g. through narrowing based off of `new.target`), so the signatures are not distinct, even though the behavior is.